requirements.txt
asgiref==3.11.0
click==8.3.1
dj-database-url==3.0.1
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.4
uvicorn==0.40.0
whitenoise==6.11.0
